pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mcstatus"
version = "0.1.3"
description = "Check Minecraft server status"
authors = ["Jigsaw <j1g5aw@foxmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Jigsaw111/nonebot_plugin_puppet"
repository = "https://github.com/Jigsaw111/nonebot_plugin_puppet"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-alpha.11"
nonebot-adapter-cqhttp = "^2.0.0a13"
nonebot-plugin-apscheduler = "^0.1.2"
mcstatus = "^5.2.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.4.1"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"